Start refactoring to be in polars-ops and use expressions.
