+++ crates/polars-core/src/chunked_array/ops/mod.rs
@@ -34,7 +34,6 @@ pub(crate) mod nulls;
 mod reverse;
 #[cfg(feature = "rolling_window")]
 pub(crate) mod rolling_window;
-mod search;
 pub mod search_sorted;
 mod set;
 mod shift;
@@ -236,12 +235,6 @@ pub trait ChunkApply<'a, T> {
     // (value of chunkedarray, value of slice) -> value of slice
     where
         F: Fn(Option<T>, &S) -> S;
-}
-
-/// Search for an item.
-pub trait ChunkSearch<'a, T> {
-    /// Return the index of the given value within self, or `None` if not found.
-    fn index_of(&'a self, value: Option<T>) -> Option<usize>;
 }
 
 /// Aggregation operations.
