+++ crates/polars-python/src/series/scatter.rs
@@ -150,64 +150,3 @@ impl PySeries {
         index_of(&self.series, &value.series).map_err(|e| PyErr::from(PyPolarsErr::from(e)))
     }
 }
-
-/// Try casting the value to the correct type, then call index_of().
-macro_rules! try_index_of {
-    ($self:expr, $value:expr) => {{
-        let cast_value = $value.map(|v| AnyValue::from(v).strict_cast($self.dtype()));
-        if cast_value == Some(None) {
-            // We can can't cast the searched-for value to a valid data point
-            // within the dtype of the Series we're searching, which means we
-            // will never find that value.
-            None
-        } else {
-            let cast_value = cast_value.flatten();
-            $self.index_of(cast_value.map(|v| v.extract().unwrap()))
-        }
-    }};
-}
-
-fn index_of(series: &Series, value_series: &Series) -> PolarsResult<Option<usize>> {
-    let value_series = if value_series.dtype().is_null() {
-        // Should be able to cast null dtype to anything, so cast it to dtype of
-        // Series we're searching.
-        &value_series.cast(series.dtype())?
-    } else {
-        value_series
-    };
-    let value_dtype = value_series.dtype();
-
-    if value_dtype.is_signed_integer() {
-        let value = value_series.cast(&DataType::Int64)?.i64().unwrap().get(0);
-        return Ok(downcast_as_macro_arg_physical!(series, try_index_of, value));
-    }
-    if value_dtype.is_unsigned_integer() {
-        let value = value_series.cast(&DataType::UInt64)?.u64().unwrap().get(0);
-        return Ok(downcast_as_macro_arg_physical!(series, try_index_of, value));
-    }
-    if value_dtype.is_float() {
-        let value = value_series.cast(&DataType::Float64)?.f64().unwrap().get(0);
-        return Ok(downcast_as_macro_arg_physical!(series, try_index_of, value));
-    }
-    // At this point we're done handling integers and floats.
-    match value_series.dtype() {
-        DataType::List(_) => {
-            let value = value_series
-                .list()
-                .unwrap()
-                .get(0)
-                .map(|arr| Series::from_arrow("".into(), arr).unwrap());
-            Ok(series.list()?.index_of(value.as_ref()))
-        },
-        #[cfg(feature="dtype-array")]
-        DataType::Array(_, _) => {
-            let value = value_series
-                .array()
-                .unwrap()
-                .get(0)
-                .map(|arr| Series::from_arrow("".into(), arr).unwrap());
-            Ok(series.array()?.index_of(value.as_ref()))
-        },
-        _ => unimplemented!("TODO"),
-    }
-}
